+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
-## 1.7.0 - Npc Generator
-- NPC Sheet : Added a new generator feature.
+## 1.7.0 - DiceRoller for Techniques & Npc Generator
+- NPC Sheet : Added generator feature (demeanor, clan and families names courteously authorized by Edge).
+- PC/NPC sheet:
+    - Added the ability to techniques, with a skill set, to open the DicePicker with presets values.
+    - Some can interact with targets, but do the default difficulty if none.
+- Compendiums :
+    - Techniques : Added difficulty and skill values.
 
 ## 1.6.0 - QoL & SoftLock
 - PC/NPC/Armies sheet: